added media functionality

+++ routes/admin.php
@@ -15,14 +15,20 @@
 Route::get('/', 'DashboardController')->name('admin.dashboard');
 
 // Projects
-Route::get('/projects/create', 'ProjectController@create');
-Route::post('/projects', 'ProjectController@store');
-Route::get('/projects', 'ProjectController@index');
-Route::get('/projects/{slug}/edit', 'ProjectController@edit');
-Route::patch('/projects/{id}', 'ProjectController@update');
-Route::delete('/projects/{id}', 'ProjectController@destroy');
+Route::get('/projects/create', 'ProjectController@create')->name('admin.projects.create');
+Route::post('/projects', 'ProjectController@store')->name('admin.projects.store');
+Route::get('/projects', 'ProjectController@index')->name('admin.projects.index');
+Route::get('/projects/{slug}/edit', 'ProjectController@edit')->name('admin.projects.edit');
+Route::patch('/projects/{id}', 'ProjectController@update')->name('admin.projects.update');
+Route::delete('/projects/{id}', 'ProjectController@destroy')->name('admin.projects.destroy');
+Route::delete('/projects/{projectId}/tags/{tagId}', 'ProjectController@deletetag')->name('admin.projects.tags.destroy');
 
-// tags
+// Media
+Route::post('/media/{slug}/create', 'ProjectMediaController@create')->name('admin.projects.media.create');
+Route::delete('/projects/{projectId}/media/{mediaId}', 'ProjectMediaController@destroy')->name('admin.projects.media.destroy');
+
+
+// Tags
 Route::get('/tags', 'TagController@index')->name('admin.tags.index');
 Route::get('/tags/create', 'TagController@create')->name('admin.tags.create');
 Route::post('/tags', 'TagController@store')->name('admin.tags.store');
